+++ docs/release_notes.rst
@@ -11,3 +11,4 @@ Enhancements
 
 - Added ``WaveShaper`` signal node operator (:pull:`100`).
 - Added ``PitchShift`` effect node (:pull:`101`).
+- Added ``FrequencyShifter`` effect node (:pull:`102`).